requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
deepdiff==6.6.1
idna==3.4
ordered-set==4.1.0
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.18.1
ruamel.yaml.clib==0.2.8
urllib3==2.0.7
